Add a new garbage state, between hanging and dropping
This state is required to keep clients synchronized.
The server move a garbage from hanging to waiting state (waiting to be
dropped). Once he receives the change, the targetted field actually
drops the garbage and inform the server back of the effective drop tick.

gb_hang_tk is no longer required to be lower than tk_lag_max. But the
actualy hang time may be greater than gb_hang_tk due to lag.