Move applyStyle() methods to a Stylable "mixin" class
applyStyle() was needed for non-widget class (e.g. field display labels).

We want to factorize property handling (e.g. various property names for a
sf::Text) but without limiting style searching (for screens, we still to to
search in fallbacks).

Now, widgets are basically stylable with automatic "Pos" processing and an
attached screen.