Improve IniFile interface using std::initializer_list
Allow to access a key using a list of path components.
Previously it was only supported for two components (section and key).

This prepares work for nicer style handling and use of dotted style names
instead of simple prefix.

Also, this allows to define a get()/set() with default parameter for functions
without the section parameter.
Accesses was updated to use (for instance)
  get("X.Y", default)
instead of
  get("X", "Y", default)